feat: add changes according to code review #P23-223
